chore: wasmer setup and cleanup

+++ fetch_sources/fetch_sources.cc
@@ -60,15 +60,13 @@ static void load_fetched_sources_yaml(
 			std::string item_str{item.val().str, item.val().len};
 			auto        inc_runfile_path = options.runfiles->Rlocation(item_str);
 
-			if(!inc_runfile_path.empty()) {
-				if(fs::exists(inc_runfile_path)) {
-					fs::copy_file(
-						inc_runfile_path,
-						dirname / fs::path{inc_runfile_path}.filename(),
-						ec
-					);
-					util::report_error_code_and_exit(options.reporter, ec);
-				}
+			if(!inc_runfile_path.empty() && fs::exists(inc_runfile_path)) {
+				fs::copy_file(
+					inc_runfile_path,
+					dirname / fs::path{inc_runfile_path}.filename(),
+					ec
+				);
+				util::report_error_code_and_exit(options.reporter, ec);
 			} else {
 				options.reporter.report(ecsact_rtb::warning_message{
 					"Unknown fetched include: " + item_str,
@@ -86,15 +84,13 @@ static void load_fetched_sources_yaml(
 		for(const auto& item : entry.children()) {
 			std::string item_str{item.val().str, item.val().len};
 			auto        inc_runfile_path = options.runfiles->Rlocation(item_str);
-			if(!inc_runfile_path.empty()) {
-				if(fs::exists(inc_runfile_path)) {
-					fs::copy_file(
-						inc_runfile_path,
-						dirname / fs::path{inc_runfile_path}.filename(),
-						ec
-					);
-					util::report_error_code_and_exit(options.reporter, ec);
-				}
+			if(!inc_runfile_path.empty() && fs::exists(inc_runfile_path)) {
+				fs::copy_file(
+					inc_runfile_path,
+					dirname / fs::path{inc_runfile_path}.filename(),
+					ec
+				);
+				util::report_error_code_and_exit(options.reporter, ec);
 			} else {
 				options.reporter.report(ecsact_rtb::warning_message{
 					"Unknown fetched source: " + item_str,
@@ -111,13 +107,11 @@ static void load_fetched_sources_yaml(
 		for(const auto& item : entry.children()) {
 			std::string item_str{item.val().str, item.val().len};
 			auto        inc_runfile_path = options.runfiles->Rlocation(item_str);
-			if(!inc_runfile_path.empty()) {
-				if(fs::exists(inc_runfile_path)) {
-					fs::copy_file(
-						inc_runfile_path,
-						dirname / fs::path{inc_runfile_path}.filename()
-					);
-				}
+			if(!inc_runfile_path.empty() && fs::exists(inc_runfile_path)) {
+				fs::copy_file(
+					inc_runfile_path,
+					dirname / fs::path{inc_runfile_path}.filename()
+				);
 			} else {
 				options.reporter.report(ecsact_rtb::warning_message{
 					"Unknown fetched share: " + item_str,
